refactor(libs/utils): created constants folder and moved constants file inside constants folder
