+++ libs/utils/src/index.ts
@@ -1,1 +1,1 @@
-export * from './constants';
+export * from './constants/constants';
